move view tests to sparql_endpoints
