+++ backend/sources/conftest.py
@@ -3,7 +3,7 @@ from items import namespace as item
 from rdf.ns import FOAF, SCHEMA
 from rdf.utils import graph_from_triples
 from rdflib import Literal
-from sparql.conftest import sparql_client, sparql_user
+from sparql_endpoints.conftest import sparql_client, sparql_user
 
 from . import namespace as my
 from .graph import graph
